refactor: 레거시 mvc 삭제
