--- mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/ForwardController.java
@@ -1,19 +1,0 @@
-package com.interface21.webmvc.servlet.mvc.asis;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.util.Objects;
-
-public class ForwardController implements Controller {
-
-    private final String path;
-
-    public ForwardController(String path) {
-        this.path = Objects.requireNonNull(path);
-    }
-
-    @Override
-    public String execute(HttpServletRequest request, HttpServletResponse response) {
-        return path;
-    }
-}
